Further modified the dvorak layout.
Note: one modification I made was swapping the j & k keys back to their original
dvorak positions. This is great for typing, but not so great for vim. I refused
to give up this luxury in vim, so I remapped the keys in vim. So in vim,
everything works as expected. But now every application outside of vim is
backward. I'm not sure what will come of this as of yet.