Came up with a new modified dvorak layout.
This layout fixes the extra stress on the right pinky by swapping the `l' and
`b' keys. The only digraphs that are negatively affected by this change are the
`gl', `fl', `lf', and `ld', but I find these digraphs to appear very infrequently in
practice.

The big win is putting a letter (b) in place of QWERTY `p' that is even rarer
than `p', along with the minimal amount of changes to the dvorak layout itself.