Changes to the dvorak layout.
Mostly turning the layout into a angle/wide layout. The picture of the new
layout looks pretty drastic, but not that much has changed. Typing like this is
so much more comfortable as Enter and Left Shift are within reach of the pinky
without having to turn the wrist at an awkward angle. I really like the new
positions for ' and /, too. We'll see about where z has gone. It has some very
prime real estate.

<C-l> is now the preferred alias for Escape in vim. <C-h> is actually extremely
useful as BackSpace (who knew??).