Mixed up some bottom row behavior
The key to the left of the SpaceBar is now mapped to Return. The key to
the right of the ghetto one is now mapped to BackSpace. This decision
was made because I use BackSpace a whole heck of a lot more than
Return (and there are alternatives to Return in a lot of my favorite
utilities like C-j and C-m in emacs and terminal).