Add small time delay in between cell makes
It seems like when jupyter tries to make a whole bunch of cells very quickly via
RPC something goes wrong. Adding a small time delay seems to fix the issue. In
the long run it would be nice to figure out exactly why this does not work. It
really seems like an RPC thing. Perhaps one thing to try would be to kill the
RPC elisp server and then restart it.

But I already confirmed that I can send a single command over RPC. It must be
due to the deluge of commands over RPC in a very short amount of time. But it's
strange because slowly building up test_api.py actually worked...